Add files via upload
This Geographical visualization of the World Covid dataset as Choropleth Map created in python using json library and json file. this clearly shows the Worldwide spread of Covid till date in terms of Cumulative cases, Which countries are most effected and have largest number of cases can be clearly spotted in darker shades. United States of America is leading with number of cases followed by India. There are some countries with no color or black color, that means there was no data for those countries in the dataset